refactor adapter
