+++ state/id.go
@@ -4,7 +4,33 @@ import (
 	"bytes"
 	"crypto/sha256"
 	"encoding/binary"
+	"math"
 )
+
+const (
+	AccountPath = "/accounts/"
+
+	DeploymentPath         = "/deployments/"
+	DeploymentSequencePath = "/deployments-seq/"
+
+	DeploymentGroupPath = "/deployment-groups/"
+	ProviderPath        = "/providers/"
+	OrderPath           = "/orders/"
+	FulfillmentPath     = "/fulfillment-orders/"
+	LeasePath           = "/lease/"
+
+	MaxRangeLimit = math.MaxInt64
+
+	AddressSize = 32 // XXX: check
+)
+
+func MaxAddress() []byte {
+	return bytes.Repeat([]byte{0xff}, AddressSize)
+}
+
+func MinAddress() []byte {
+	return make([]byte, AddressSize)
+}
 
 func DeploymentAddress(account []byte, nonce uint64) []byte {
 	return NonceAddress(account, nonce)
